create movie slice and and to redux and create a custom hook
