+++ netflix-gpt-swati/src/components/Browse.js
@@ -1,19 +1,10 @@
 import Header from './Header'
-import React, { useEffect } from 'react';
-import { API_OPTIONS } from './utils/constants';
+import React from 'react';
+import useNowPlayingMovies from '../hooks/useNowPlayingMovies';
 
 const Browse = () => {
 
-  const getNowPlayingMovie = async() => {
-    const data = await fetch('https://api.themoviedb.org/3/movie/now_playing?language=en-US&page=1', API_OPTIONS);
-    const json = await data.json();
-    console.log(json.results)
-  }
-
-  useEffect(()=>{
-    getNowPlayingMovie();
-  },[])
-
+  useNowPlayingMovies();
   return (
     <div><Header/></div>
   )
