+++ netflix-gpt-swati/src/components/utils/moviesSlice.js
@@ -1,0 +1,16 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const moviesSlice = createSlice({
+    name: 'movies',
+    initialState: {
+        nowPlayingMovies: null,
+    },
+    reducers: {
+        addNowPlayingMovies: (state, action) => {
+            state.nowPlayingMovies = action.payload;
+        },
+    },
+});
+
+export const { addNowPlayingMovies } = moviesSlice.actions;
+export default moviesSlice.reducer;
